+++ changelog.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * Patch inventory item request.
     * Delete inventory item request.
     * Patch user request.
+    * Delete user request.
     
 ### Changed